+++ src/MapboxPlugin.php
@@ -15,12 +15,14 @@ use Craft;
 use craft\base\Field;
 use craft\base\Model;
 use craft\base\Plugin;
+use craft\events\ConfigEvent;
 use craft\events\ModelEvent;
 use craft\events\PluginEvent;
 use craft\events\RegisterComponentTypesEvent;
 use craft\helpers\UrlHelper;
 use craft\services\Fields;
 use craft\services\Plugins;
+use craft\services\ProjectConfig;
 use doublesecretagency\mapbox\fields\AddressField;
 use doublesecretagency\mapbox\models\Settings;
 use doublesecretagency\mapbox\web\twig\Extension;
@@ -61,6 +63,9 @@ class MapboxPlugin extends Plugin
 
         // Register all events
         $this->_registerFieldType();
+
+        // Manage conversions of the Address field
+        $this->_manageFieldTypeConversions();
 
         // Redirect after plugin install
         $this->_postInstallRedirect();
@@ -109,6 +114,84 @@ class MapboxPlugin extends Plugin
             Fields::EVENT_REGISTER_FIELD_TYPES,
             static function (RegisterComponentTypesEvent $event) {
                 $event->types[] = AddressField::class;
+            }
+        );
+    }
+
+    /**
+     * Manage conversions of the Address field from the Google Maps plugin.
+     *
+     * @return void
+     */
+    private function _manageFieldTypeConversions(): void
+    {
+        // If Google Maps plugin is not installed, bail
+        if (!Craft::$app->getPlugins()->isPluginEnabled('google-maps')) {
+            return;
+        }
+
+        // When a single project config line gets updated
+        Event::on(
+            ProjectConfig::class,
+            ProjectConfig::EVENT_UPDATE_ITEM,
+            static function (ConfigEvent $event) {
+
+                // Get old and new types
+                $oldType = $event->oldValue['type'] ?? null;
+                $newType = $event->newValue['type'] ?? null;
+
+                // If old type wasn't a Google Maps Address field, bail
+                if (!($oldType === 'doublesecretagency\googlemaps\fields\AddressField')) {
+                    return;
+                }
+
+                // If new type is not a Mapbox Address field, bail
+                if (!($newType === 'doublesecretagency\mapbox\fields\AddressField')) {
+                    return;
+                }
+
+                // Get the field's UID
+                $uid = str_replace('fields.', '', $event->path);
+
+                // Get the actual field
+                $field = Craft::$app->getFields()->getFieldByUid($uid);
+
+                // If unable to get the field, bail
+                if (!$field) {
+                    return;
+                }
+
+                // List of columns to copy between tables
+                $columns = [
+                    'elementId', 'fieldId',
+                    'formatted', 'raw',
+                    'name', 'street1', 'street2',
+                    'city', 'state', 'zip',
+                    'county', 'country',
+                    'lng', 'lat', 'zoom',
+                    'dateCreated', 'dateUpdated', 'uid'
+                ];
+
+                // Merge and escape column names
+                $columns = '[['.implode(']],[[', $columns).']]';
+
+                // Copy field's rows from `googlemaps_addresses` into `mapbox_addresses`
+                $sql = "
+INSERT INTO [[mapbox_addresses]] ({$columns})
+SELECT {$columns}
+FROM [[googlemaps_addresses]]
+WHERE [[fieldId]] = :fieldId
+  AND NOT EXISTS (
+    SELECT 1
+    FROM [[mapbox_addresses]]
+    WHERE [[mapbox_addresses]].[[uid]] = [[googlemaps_addresses]].[[uid]]
+)";
+
+                // Execute the SQL statement
+                \Yii::$app->db->createCommand($sql)
+                    ->bindValues([':fieldId' => $field->id])
+                    ->execute();
+
             }
         );
     }